% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md2s-package.R
\docType{package}
\name{md2s-package}
\alias{md2s-package}
\title{Multi-Dataset Multidimensional Scaling}
\description{
The MD2S package computes a latent factor common to two matrices as well
as an idiosyncratic factor unique to each. In addition, it offers a flexible
modeling strategy that permits the scaled locations to be a function of
covariates, and efficient implementation allows for inference through
resampling
}
\author{
Ted Enamorado \email{ted.enamorado@gmail.com},

Maintainer: Peter Bachman \email{bachman.p@wustl.edu}
}
\keyword{package}
