% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md2s.R
\name{md2s}
\alias{md2s}
\title{md2s}
\usage{
md2s(
  X,
  y,
  X.s = NULL,
  X.X = NULL,
  X.y = NULL,
  init = "svd",
  dim = 1,
  tol = 1e-06,
  BIC = FALSE
)
}
\arguments{
\item{X}{First matrix of realized outcomes \eqn{m = 1} for \eqn{N} observations and \eqn{K_1} covariates; dimensions \eqn{N \times K_1}.}

\item{y}{Second matrix of realized outcomes \eqn{m = 2} for \eqn{N} observations and \eqn{K_2} covariates; dimensions \eqn{N \times K_1}}

\item{X.s}{Optional matrix of covariates for estimating scaled locations in the shared subspace.}

\item{X.X}{Optional matrix of covariates for estimating scaled locations in the idiosyncratic subspace of \code{X}.}

\item{X.y}{Optional matrix of covariates for estimating scaled locations in the idiosyncratic subspace of \code{y}.}

\item{init}{INOPERABLE. Matrix initialization method.}

\item{dim}{Number of dimensions to be estimated by the model (from most- to least-explanatory).}

\item{tol}{Iteration tolerance threshold. Criteria for termination of iteration.}

\item{BIC}{INOPERABLE. Bayesian Information Criterion.}
}
\value{
An object containing:
\item{z}{TODO}
\item{z.X}{TODO}
\item{z.y}{TODO}
\item{w.Xs}{TODO}
\item{w.ys}{TODO}
\item{beta.z}{TODO}
\item{proportionX}{TODO}
}
\description{
Multi-Dataset Multidimensional Scaling
}
\examples{
\dontrun{
md2s()
}




}
\seealso{
\code{\link[=md2sPermute]{md2sPermute()}}
}
\author{
Peter Bachman \href{mailto:bachman.p@wustl.edu}{bachman.p@wustl.edu}, Patrick Edwards \href{mailto:edwards.p@wustl.edu}{edwards.p@wustl.edu}, and Zion Little \href{mailto:l.zion@wustl.edu}{l.zion@wustl.edu}
}
