% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MD2Spermute.R
\name{md2Spermute}
\alias{md2Spermute}
\alias{md2spermute}
\title{Permutation Test}
\usage{
md2Spermute(kX.num, n, ky, nsims, nperm, nboot)
}
\arguments{
\item{kX.num}{K(2) as defined in the paper.}

\item{n}{Number of rows in the matrix}

\item{ky}{K(1) in the paper}

\item{nsims}{The number of simulations to be done. Default is set to 200.}

\item{nperm}{The number of permutations to be done. Default is set to 200.}

\item{nboot}{The number of bootstraps to be performed. Default is set to 200.}

\item{d}{number of dimensions}

\item{fpath}{Designated file path where we will store the results}
}
\value{
An object containing:
\item \code{kX.num} {Number of columns in the matrix, K(2) in the paper}
\item \code{n} {Number of rows in the matrix}
\item \code{ky} {Number of columns of matrix, K(1) in the paper}
\item \code{nsims} {Number of simulations done}
\item \code{nperm} {Number of permutations done}
\item \code{nboot} {Number of bootstraps done}
\item \code{fpath} {Designated file path for the result}
}
\description{
Permutation Test
}
\examples{
md2Spermute(kX.num = 100, n = 50, ky = 40, nsims = 200, nperm = 200, nboot = 200)
}
\seealso{
\code{\link[MD2S]{MD2S}}
}
\author{
Cecilia Y. Sui and Evan E. Jo: \email{c.sui@wustl.edu} \email{ejo@wustl.edu}
}
