% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md2s-package.R
\docType{package}
\name{md2s-package}
\alias{md2s}
\alias{md2s-package}
\title{md2s: Multi-Dataset Multidimensional Scaling}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Peter Bachman \email{bachman.p@wustl.edu}

Authors:
\itemize{
  \item Ted Enamorado \email{ted@wustl.edu}
  \item Gabriel López-Moctezuma \email{glmoctezuma@caltech.edu}
  \item Marc Ratkovic \email{ratkovic@princeton.edu}
}

Other contributors:
\itemize{
  \item Patrick Edwards \email{edwards.p@wustl.edu} [contributor]
  \item Zion Little \email{l.zion@wustl.edu} [contributor]
  \item Cecilia Sui \email{c.sui@wustl.edu} [contributor]
  \item Evan Jo \email{ejo@wustl.edu} [contributor]
}

}
\keyword{internal}
