% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md2s.R
\name{md2s}
\alias{md2s}
\alias{MD2S}
\title{md2s}
\usage{
md2s(
  X,
  y,
  X.s = NULL,
  X.X = NULL,
  X.y = NULL,
  init = "svd",
  dim = 1,
  sim = FALSE,
  tol = 1e-06,
  BIC = FALSE
)
}
\arguments{
\item{X}{TODO}

\item{y}{TODO}

\item{X.s}{TODO}

\item{X.X}{TODO}

\item{X.y}{TODO}

\item{init}{TODO}

\item{dim}{Number of dimensions}

\item{sim}{TODO}

\item{tol}{TODO}

\item{BIC}{TODO}
}
\value{
An object containing:
\item{z}{TODO}
\item{z.X}{TODO}
\item{z.y}{TODO}
\item{w.Xs}{TODO}
\item{w.ys}{TODO}
\item{beta.z}{TODO}
\item{proportionX}{TODO}
}
\description{
Multi-Dataset Multidimensional Scaling
}
\examples{
\dontrun{
md2s()
}

}
\seealso{
\code{\link[md2sPermute]{md2sPermute}}
}
\author{
Peter Bachman \href{mailto:bachman.p@wustl.edu}{bachman.p@wustl.edu}, Patrick Edwards \href{mailto:edwards.p@wustl.edu}{edwards.p@wustl.edu}, and Zion Little \href{mailto:l.zion@wustl.edu}{l.zion@wustl.edu}
}
