% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md2sPermute.R
\name{md2sPermute}
\alias{md2sPermute}
\alias{md2spermute}
\title{Permutation Test}
\usage{
md2sPermute(
  kX.num = 100,
  n = 50,
  ky = 40,
  nsims,
  nperm,
  nboot,
  fpath = ".",
  sim = TRUE,
  d = 5
)
}
\arguments{
\item{kX.num}{K(2) as defined in the paper.}

\item{n}{Number of rows in the matrix}

\item{ky}{K(1) in the paper}

\item{nsims}{The number of simulations to be done. Default is set to 200.}

\item{nperm}{The number of permutations to be done. Default is set to 200.}

\item{nboot}{The number of bootstraps to be performed. Default is set to 200.}

\item{fpath}{Designated file path where we will store the results}

\item{d}{number of dimensions}
}
\value{
An object containing:
\item{kX.num}{Number of columns in the matrix, K(2) in the paper}
\item{n}{Number of rows in the matrix}
\item{ky}{Number of columns of matrix, K(1) in the paper}
\item{nsims}{Number of simulations done}
\item{nperm}{Number of permutations done}
\item{nboot}{Number of bootstraps done}
\item{fpath}{Designated file path for the result}
}
\description{
Permutation Test
}
\examples{
\dontrun{
md2sPermute(kX.num = 100, n = 50, ky = 40, nsims = 200, nperm = 200, nboot = 200)
}





}
\seealso{
\code{\link[=md2s]{md2s()}}
}
\author{
Cecilia Y. Sui and Evan E. Jo: \email{c.sui@wustl.edu} \email{ejo@wustl.edu}
}
